New Mana Bar and time stop uses mana now

+++ src/main/java/net/skyeshade/wol/client/ClientStatsData.java
@@ -1,41 +1,41 @@
 package net.skyeshade.wol.client;
 
 public class ClientStatsData {
-    private static int playerMana;
-    private static int playerMaxMana;
-    private static int playerManaCore;
-    private static int playerManaCoreExhaustion;
-    private static int playerMaxManaCore;
+    private static long playerMana;
+    private static long playerMaxMana;
+    private static long playerManaCore;
+    private static long playerManaCoreExhaustion;
+    private static long playerMaxManaCore;
 
-    public static void setMana(int mana) {
+    public static void setMana(long mana) {
         ClientStatsData.playerMana = mana;
     }
-    public static void setMaxMana(int max_mana) {
+    public static void setMaxMana(long max_mana) {
         ClientStatsData.playerMaxMana = max_mana;
     }
-    public static int getPlayerMana() {
+    public static long getPlayerMana() {
         return playerMana;
     }
-    public static int getPlayerMaxMana() {
+    public static long getPlayerMaxMana() {
         return playerMaxMana;
     }
 
-    public static void setManaCore(int manacore) {
+    public static void setManaCore(long manacore) {
         ClientStatsData.playerManaCore = manacore;
     }
-    public static void setMaxManaCore(int max_manacore) {
+    public static void setMaxManaCore(long max_manacore) {
         ClientStatsData.playerMaxManaCore = max_manacore;
     }
-    public static void setManaCoreExhaustion(int manacore_exhaustion) {
+    public static void setManaCoreExhaustion(long manacore_exhaustion) {
         ClientStatsData.playerManaCoreExhaustion = manacore_exhaustion;
     }
-    public static int getPlayerManaCore() {
+    public static long getPlayerManaCore() {
         return playerManaCore;
     }
-    public static int getPlayerMaxManaCore() {
+    public static long getPlayerMaxManaCore() {
         return playerMaxManaCore;
     }
-    public static int getPlayerManaCoreExhaustion() {
+    public static long getPlayerManaCoreExhaustion() {
         return playerManaCoreExhaustion;
     }
 }
