+++ src/main/java/net/skyeshade/wol/networking/packet/mana/MaxManaDataSyncS2CPacket.java
@@ -8,18 +8,18 @@ import net.skyeshade.wol.client.ClientStatsData;
 import java.util.function.Supplier;
 
 public class MaxManaDataSyncS2CPacket {
-    private final int max_mana;
+    private final long max_mana;
 
-    public MaxManaDataSyncS2CPacket(int max_mana) {
+    public MaxManaDataSyncS2CPacket(long max_mana) {
         this.max_mana = max_mana;
     }
 
     public MaxManaDataSyncS2CPacket(FriendlyByteBuf buf) {
-        this.max_mana = buf.readInt();
+        this.max_mana = buf.readLong();
     }
 
     public void toBytes(FriendlyByteBuf buf) {
-        buf.writeInt(max_mana);
+        buf.writeLong(max_mana);
     }
 
     public boolean handle(Supplier<NetworkEvent.Context> supplier) {
